refactor: separate pdf reading from openai logic
